backend/requirements.txt
# CredoCarbon Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.3.2
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities - Pydantic 2.4.2 has manylinux wheels
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-multipart==0.0.6

# File handling
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2

# Date handling
python-dateutil==2.8.2

# PDF generation
reportlab==4.0.7

# Excel generation
openpyxl==3.1.2

# Document generation
python-docx==1.1.0

# Google Cloud Storage
google-cloud-storage==2.14.0

# Google Cloud Pub/Sub (for event bus)
google-cloud-pubsub==2.18.4

# Google Cloud Tasks (for task queue)
google-cloud-tasks==2.14.2
